[Compose] Refactor out MotionLayout.kt classes
This refactors the following classes into separate files (with their
respective interfaces, helper functions):
- Transition.kt: Transition, (Composable) Transition, TransitionImpl
- MotionScene.kt: MotionScene, (Composable) MotionScene, JSONMotionScene
- MotionMeasurer.kt: MotionMeasurer

Applied formatting rest of package
